+++ app/src/provider/AllProvider.tsx
@@ -1,0 +1,30 @@
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { LocationProvider } from '../context/location/Location.tsx'
+import { UnitsProvider } from '../context/unitsSystem/UnitsSystem.tsx'
+import { FavoritesProvider } from '../context/favoritesLocation/FavoritesContext.tsx'
+import type { ReactNode } from 'react'
+
+const queryClient = new QueryClient({
+    defaultOptions: {
+        queries: {
+            retry: 3,
+            refetchOnWindowFocus: false,
+            refetchInterval: 300000,
+        }
+    }
+});
+
+export default function AllProvider({ children }: { children: ReactNode }) {
+
+    return (
+        <QueryClientProvider client={queryClient}>
+            <FavoritesProvider>
+                <UnitsProvider>
+                    <LocationProvider>
+                        {children}
+                    </LocationProvider>
+                </UnitsProvider>
+            </FavoritesProvider>
+        </QueryClientProvider>
+    )
+}